+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## TODO
 - 
 
+## Release 2024-03-15
+- Fix name of the 7xHE10-17, 7xHE18-22 buckets
+
 ## Release 2024-03-13
 - Rename lastBusinessDayPrior() to lastBusinessDayBefore()
 - Added solar buckets 5xHE10-17, 5xHE18-22, 7xHE10-17, 7xHE18-22,